Introducing course member concept
