+++ lib/mumuki/classroom/models.rb
@@ -1,6 +1,7 @@
 require_relative './models/concerns/extensions'
 require_relative './models/concerns/with_submission_process'
 require_relative './models/concerns/with_failed_submission_reprocess'
+require_relative './models/concerns/course_member'
 
 require_relative './models/document'
 require_relative './models/message'
